update
Fixed get_topics function.
Merged small functions with other files.
